requirements.txt
Scrapy
requests
Pillow
Protego
lxml
twisted